requirements.txt
selenium
webdriver_manager
dotenv-python
